Update get pizza order details
This piece of code gathers information from the user regarding the size of the pizza they want which can be Small Medium, Large or X Large. It calculates the price based on the selected size and asks for the number of pizzas to be ordered. After that, it creates a Pizza object that includes all the collected information. This allows users to easily place their pizza orders.